reorganized structure, added new script to compile
