+++ README.md
@@ -1,2 +1,8 @@
 # ECSE429Project
 Final Project for ECSE 429, Software Validation at McGill University, Fall of 2019
+
+## Authors:
+* Pierre Robert-Michon
+* Romain Couperier
+* Paul Seimandi
+* Milo Sobral
