+++ src/tools/compile_and_run.py
@@ -1,0 +1,2 @@
+# Script providing functions to compile and run C programs from a python script
+# It is iportant that the input and output of thos scripts can be specified
